Track the inferior process explicitly
We assume at any point there is ONE inferior process which is tracking our buffer.
We keep a reference to the buffer as soon as it is created.

(python-shell-get-process) does a lot of work to re-compute a buffer name, and
actually returns nil if the inferior has died. This prevents us to track the
exit status correctly.

In the future, this would allow us to launch a new python process and attach to
an existing inferior.